Programming Assignment 2
In this programming assignment, the Geometric object class is provided for me. I design a UML diagram for the Triangle class and implement the classes. This test program prompts the user to enter three sides of the triangle, a color, and a Boolean value to indicate whether the triangle is filled.